Support GHC 8/template haskell 2.11.
This needed a new way to compute the required signatures for GADT lenses.
Previously these were computed from the equality constraints that were
on the data type definition that was the template haskell representation
of GADTs. Now, there's two extra constructors for GADTs, so we have to
generate the substitutions from the list of type variables in the data
definition, and the result type of the constructor.

Additionally, for the totality check the indices are now on the Gadt
constructor instead of in an equality constraint. So we skip the Foralls
(these contain only explicit equality constraints or type/kind variables)
and instead unpack the result type of the GADT.

This also adds an extra test case to make sure this works for lists of
type variables with length > 1.